Start breaking off rightsMetadata stuff from Publishable
Introducing Rightsable.  Possible alternative name "Endowable".
