+++ lib/dor/models/rightsable.rb
@@ -1,0 +1,25 @@
+module Dor
+  module Rightsable
+    extend ActiveSupport::Concern
+
+    included do
+      has_metadata :name => 'rightsMetadata', :type => Dor::RightsMetadataDS, :label => 'Rights metadata'
+    end
+
+    def build_rightsMetadata_datastream(ds)
+      content_ds = self.admin_policy_object.datastreams['defaultObjectRights']
+      ds.dsLabel = 'Rights Metadata'
+      ds.ng_xml = content_ds.ng_xml.clone
+      ds.content = ds.ng_xml.to_xml
+    end
+
+    def world_doc
+      return Nokogiri::XML::Builder.new do |xml|
+        xml.access(:type => 'read') {
+          xml.machine { xml.world }
+        }
+      end.doc
+    end
+
+  end
+end
